- move test ensembler implementation outside of test module
